Coupon test mixin
