+++ ecommerce/extensions/test/factories.py
@@ -1,12 +1,5 @@
 from oscar.test.factories import *  # pylint:disable=wildcard-import,unused-wildcard-import
 
-from ecommerce.extensions.api.v2.views.coupons import CouponViewSet
-from ecommerce.tests.factories import PartnerFactory
-
-Benefit = get_model('offer', 'Benefit')
-Catalog = get_model('catalogue', 'Catalog')
-Voucher = get_model('voucher', 'Voucher')
-
 OrderNumberGenerator = get_class('order.utils', 'OrderNumberGenerator')
 
 
@@ -45,42 +38,3 @@ def create_order(number=None, basket=None, user=None, shipping_address=None,  # 
         **kwargs)
     basket.set_as_submitted()
     return order
-
-
-def create_coupon(
-        title='Test coupon',
-        price=100,
-        partner=None,
-        catalog=None,
-        code='',
-        benefit_value=100,
-        sub_category=''
-    ):
-    """Helper method for creating a coupon."""
-    if partner is None:
-        partner = PartnerFactory(name='Tester')
-    if catalog is None:
-        catalog = Catalog.objects.create(partner=partner)
-    quantity = 5
-    if code is not '':
-        quantity = 1
-    data = {
-        'partner': partner,
-        'benefit_type': Benefit.PERCENTAGE,
-        'benefit_value': benefit_value,
-        'catalog': catalog,
-        'end_date': datetime.date(2020, 1, 1),
-        'code': code,
-        'quantity': quantity,
-        'start_date': datetime.date(2015, 1, 1),
-        'voucher_type': Voucher.SINGLE_USE,
-        'category': 'Test category',
-        'sub_category': sub_category
-    }
-
-    coupon = CouponViewSet().create_coupon_product(
-        title=title,
-        price=price,
-        data=data
-    )
-    return coupon
